move provider logic into useAuth hook
